+++ changelog.md
@@ -1,6 +1,7 @@
 ## 15.04.2023, Sat
 
 - 00:10 - Frontend: created "Add new bookmark" page.
+- 00:50 - Front + Back: wired up to create new bookmarks via web page.
 
 ## 14.04.2023, Fri